Adds a KeyboardAvoidingView to automatically adjust padding/height when keyboard covers content
The KeyboardAvoidingView (built-in for react-native) implements the standard behaviour of reducing the view’s height to fit above the keyboard when it is visible. Since the children views are on a ScrollView, this enables the user to scroll up and down, accessing all of the content.

On Android it defaults to no action (unnecessary)